+++ components/modules/accordion.php
@@ -1,13 +1,18 @@
 <?php
 /**
  * ACF Module: Accordian
+ *
+ * @global $data
+ * @global $row_id
  */
 
-$panels     = get_sub_field( 'accordion' );
-$open_first = get_sub_field( 'open_first' );
+use AD\App\Fields\ACF;
+
+$panels     = ACF::getRowsLayout( 'accordion', $data );
+$open_first = ACF::getField( 'open_first', $data );
 ?>
 
-<div class="panel-group" id="<?php echo $row_id; ?>" role="tablist">
+<div class="panel-group" role="tablist">
     <?php foreach ( $panels as $index => $panel ) : ?>
         <div class="panel panel-default">
             <div class="panel-heading" role="tab">
@@ -19,10 +24,10 @@ $open_first = get_sub_field( 'open_first' );
                 </h4>
             </div>
             <div id="<?php echo "{$row_id}-{$index}"; ?>"
-                 class="panel-collapse collapse <?php echo ( $open_first and $index === 0 ) ? 'in' : ''; ?>"
+                 class="panel-collapse collapse <?php echo ( $open_first && 0 === $index ) ? 'in' : ''; ?>"
                  role="tabpanel">
                 <div class="panel-body">
-                    <?php echo $panel['accordion_content']; ?>
+                    <?php echo apply_filters( 'the_content', $panel['accordion_content'] ); ?>
                 </div>
             </div>
         </div>
